requirements.txt
yt-dlp
ffmpeg-python
imagehash
pillow
openai
numpy
requests
tqdm
torch
transformers
scikit-learn
